% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardrails.R
\name{detect_cna_levels}
\alias{detect_cna_levels}
\title{Detect an accidental literal "NA" level}
\usage{
detect_cna_levels(x)
}
\arguments{
\item{x}{Character vector.}
}
\description{
Detect an accidental literal "NA" level
}
