% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{need_pkg}
\alias{need_pkg}
\title{Require an optional package with an actionable message}
\usage{
need_pkg(pkg, why)
}
\description{
Require an optional package with an actionable message
}
\keyword{internal}
