% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{perf_cindex}
\alias{perf_cindex}
\title{Out-of-sample concordance index (Harrell's C)}
\usage{
perf_cindex(obs, pred, time = NULL, event = NULL)
}
\description{
Out-of-sample concordance index (Harrell's C)
}
\keyword{internal}
