% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-attribution.R
\name{plot_attribution_plate}
\alias{plot_attribution_plate}
\title{Compose the attribution figure set into one plate}
\usage{
plot_attribution_plate(x, base_size = 9)
}
\arguments{
\item{x}{A `recoverR_attribution` object.}

\item{base_size}{Base font size.}
}
\description{
Compose the attribution figure set into one plate
}
