% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-attribution.R
\name{plot_forward_selection}
\alias{plot_forward_selection}
\title{Forward-selection staircase}
\usage{
plot_forward_selection(x, base_size = 9)
}
\arguments{
\item{x}{A `recoverR_attribution` object.}

\item{base_size}{Base font size in points.}
}
\description{
Shows the marginal gain of each added view given the views already chosen. A
staircase that flattens immediately is the honest picture of a redundant view
set.
}
