% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{reliability_curve}
\alias{reliability_curve}
\title{Calibration (reliability) curve and expected calibration error}
\usage{
reliability_curve(fit, bins = 10L)
}
\arguments{
\item{fit}{A `recoverR_conformal` object.}

\item{bins}{Number of confidence bins.}
}
\value{
A list with the binned curve and the `ece`.
}
\description{
Calibration (reliability) curve and expected calibration error
}
