% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{rr_scale_div}
\alias{rr_scale_div}
\title{Diverging scale for signed quantities}
\usage{
rr_scale_div(midpoint = 0, ...)
}
\arguments{
\item{midpoint}{Value mapped to the neutral midpoint of the diverging scale.}

\item{...}{Passed through to the learner.}
}
\description{
Diverging scale for signed quantities
}
