% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{rr_scale_seq}
\alias{rr_scale_seq}
\title{Sequential scale for ordered categories}
\usage{
rr_scale_seq(direction = 1, ...)
}
\arguments{
\item{direction}{1 for dark-to-light, -1 to reverse.}

\item{...}{Passed through to the learner.}
}
\description{
Sequential scale for ordered categories
}
